requirements.txt
python-docx==1.1.2
python-magic==0.4.27
streamlit
litellm
fastapi
uvicorn
pydantic
python-dotenv
qdrant-client
sentence-transformers
request
pandas
aiohttp>=3.8.0
boto3>=1.35.0
sse-starlette==2.1.3


